requirements.txt
fastapi
uvicorn
requests
pysocks
stem
beautifulsoup4
selenium
python-dotenv
spacy
transformers
torch
sentence-transformers
bertopic
pytesseract
neo4j
python-magic
python-slugify
pyyaml
